microservices/movie-service/web/requirements.txt
bcrypt==4.1.2
boto3==1.34.44
cryptography==42.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Jsonpify==1.5.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
matplotlib==3.8.3
pre-commit==3.6.2
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
requests
pika